requirements.txt
apache-airflow>=2.8.0,<2.9
pendulum
python-dotenv
